IDE and compilation support for each module
With this improvement, each module now compiles as a separate project
(to verify the TypeScript is valid), and supports IDE completion etc.,
because each module now declares its dependencies correctly.

To achieve this with the current tool versions, a workaround is needed
in the application level tsconfig to forced use of a single copy of
Angular and other packages across all modules.